requirements.txt
python-telegram-bot == 20.7
boto3 == 1.34.5